* add dockerfile * set default value for 'ReportRealPath' to True * move deleteSong logic from subsonic apis to native-apis
