+++ docker-compose.yml
@@ -1,0 +1,18 @@
+version: '3.4'
+services:
+  navidrome:
+    image: navidrome:latest
+    restart: unless-stopped
+    container_name: "${NAME}"
+    user: "501:20"
+    ports:
+      - "${PORT}:4533"
+    environment:
+      ND_SCANSCHEDULE: 0
+      ND_LOGLEVEL: debug
+      ND_SESSIONTIMEOUT: 24h
+      # must add this so navidrome can respond to the path navidrome is to be accessed on via traefik
+      #ND_BASEURL: /navidrome
+    volumes:
+      - "/Users/lukepurnell/subbox/${USER}/navidrome/data:/data"
+      - "/Users/lukepurnell/subbox/${USER}/nav_music:/music:ro"
